refactor: Migrate login template in a Gradle module.
