+++ app/build.gradle.kts
@@ -81,6 +81,7 @@ dependencies {
     implementation(project(":templates:onboarding"))
     implementation(project(":templates:paymentcard"))
     implementation(project(":templates:profile"))
+    implementation(project(":templates:login"))
     implementation(project(":components:fab"))
     implementation(project(":components:charts"))
     implementation(project(":components:tags"))
