+++ settings.gradle.kts
@@ -26,3 +26,4 @@ include(":components:comingsoon")
 include(":templates:onboarding")
 include(":templates:paymentcard")
 include(":templates:profile")
+include(":templates:login")
